Added email checking feature - now it is possible to register only if given email address matches the address on which was sent the verification message

+++ src/Controller/RegistrationController.php
@@ -6,6 +6,7 @@ use App\Entity\RegisterRequest;
 use App\Entity\User;
 use App\Form\RegisterRequestType;
 use App\Form\RegistrationFormType;
+use App\Services\GetUsersListService;
 use App\Services\UserRegistrationService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -13,11 +14,12 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 class RegistrationController extends AbstractController
 {
     #[Route('/register', name: 'app_register')]
-    public function registrationRequest(Request $request, EntityManagerInterface $entityManager): Response
+    public function registrationRequest(Request $request, EntityManagerInterface $entityManager, GetUsersListService $getUsersListService): Response
     {
         $session = $request->getSession();
         $hash = sha1(time());
@@ -26,9 +28,17 @@ class RegistrationController extends AbstractController
         $form = $this->createForm(RegisterRequestType::class, $registerRequest);
         $form->handleRequest($request);
         if($form->isSubmitted() && $form->isValid()) {
-            $registerRequest->setEmail(
-              $form->get('email')->getData()
-            );
+            $registeredEmailAddresses = $getUsersListService->getRegisteredEmails();
+            $email = $form->get('email')->getData();
+            if(in_array($email, $registeredEmailAddresses))
+            {
+                $message = "Na ten adres e-mail już zostało utworzone konto lub zostało wysłane zapytanie";
+                return $this->render('registration/register_request.html.twig', [
+                'message' => $message,
+                'form' => $form
+            ]);
+            }
+            $registerRequest->setEmail($email);
             $registerRequest->setHash($hash);
             $entityManager->persist($registerRequest);
             $entityManager->flush();
@@ -42,34 +52,36 @@ class RegistrationController extends AbstractController
     public function register(Request $request, UserPasswordHasherInterface $userPasswordHasher, EntityManagerInterface $entityManager, string $commonHash, UserRegistrationService $userRegistrationService): Response
     {
         $registerRequest = $entityManager->getRepository(RegisterRequest::class)->findOneBy(['hash' => $commonHash]);
+
         $session = $request->getSession();
         $hash = $session->get('hash');
-            if($hash !== $commonHash || !$registerRequest) {
-                return $this->redirectToroute('app_access_denied');
+        if($hash !== $commonHash || !$registerRequest) {
+            return $this->redirectToroute('app_access_denied');
+        }
+        $user = new User();
+        $form = $this->createForm(RegistrationFormType::class, $user);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $email = $form->get('email')->getData();
+            if($email!==$registerRequest->getEmail())
+            {
+                return $this->render('error_page/index.html.twig',[
+                    'message' => 'Ten adres e-mail nie istnieje w systemie'
+                ]);
             }
-            $user = new User();
-            $form = $this->createForm(RegistrationFormType::class, $user);
-            $form->handleRequest($request);
+            $password = $form->get('plainPassword')->getData();
+            $user->setPassword($userPasswordHasher->hashPassword($user, $password));
 
-            if ($form->isSubmitted() && $form->isValid()) {
-                $user->setPassword(
-                    $userPasswordHasher->hashPassword(
-                        $user,
-                        $form->get('plainPassword')->getData()
-                    )
-                );
+            $entityManager->persist($user);
+            $entityManager->flush();
 
-                $entityManager->persist($user);
-                $entityManager->flush();
-
-                $userRegistrationService->deleteRegisterRequest($registerRequest);
-                return $this->redirectToRoute('app_catalogs_home');
-            }
-            return $this->render('registration/register.html.twig', [
-                'registrationForm' => $form->createView(),
-            ]);
-
-
+            $userRegistrationService->deleteRegisterRequest($registerRequest);
+            return $this->redirectToRoute('app_catalogs_home');
+        }
+        return $this->render('registration/register.html.twig', [
+            'registrationForm' => $form->createView(),
+        ]);
     }
 
 }
